requirements.txt
uvicorn
fastapi
pymongo[srv]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
httpx
redis